backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pyjwt==2.8.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
paramiko==3.3.1
alembic==1.12.0
SQLAlchemy==2.0.27
